Test: 식재료 목록 조회

+++ back/src/main/java/com/gamul/api/request/IngredientListReq.java
@@ -13,7 +13,7 @@ public class IngredientListReq {
     int orderType;
 
     @JsonProperty("high_class_id")
-    Long highClassId;
+    int highClassId;
 
     @JsonProperty("user_name")
     String userName;
